Split the dataloader into a package
dataloader.py had grown to hold manifest indexing, sample/window logic,
the batching loader, the subprocess feeder, and mimi encoding. One module
per concern: manifest.py, types.py, audio.py, loader.py, subproc.py,
encode.py, with the public names re-exported from the package so callers
are unchanged.
