+++ training/dataloader/__init__.py
@@ -1,0 +1,20 @@
+"""Data loading: manifest access, sampling, batching, subprocess feeding."""
+
+from training.dataloader.audio import _load_window
+from training.dataloader.encode import encode_batch
+from training.dataloader.loader import DataLoader, _prefetch
+from training.dataloader.manifest import LazyEntries, load_entries
+from training.dataloader.subproc import SubprocessDataLoader
+from training.dataloader.types import Batch, Entry
+
+__all__ = [
+    "Batch",
+    "DataLoader",
+    "Entry",
+    "LazyEntries",
+    "SubprocessDataLoader",
+    "_load_window",
+    "_prefetch",
+    "encode_batch",
+    "load_entries",
+]
